Cleanup and add documentation

+++ atana/src/main/kotlin/nl/utwente/axini/atana/service/LastStepGroupingAndAnalysisServiceImpl.kt
@@ -7,17 +7,17 @@ import org.springframework.stereotype.Service
 
 @Service
 class LastStepGroupingAndAnalysisServiceImpl : GroupingAndAnalysisService() {
-	override fun isConfigured(): Boolean {
-		return true
-	}
+    override fun isConfigured(): Boolean {
+        return true
+    }
 
-	override fun submitAllData(model: TestModel, passingTests: Array<TestCase>, failingTests: Array<TestCase>, coverageInformation: List<TestModel>) {
-		//Do nothing
-	}
+    override fun submitAllData(model: TestModel, passingTests: Array<TestCase>, failingTests: Array<TestCase>, coverageInformation: List<TestModel>) {
+        //Do nothing
+    }
 
-	override fun submitTest(run: TestCase): AnalysisResult {
-		val problematicStep = run.steps.sortedBy { it.stepNumber.toInt() }.last()
-		return AnalysisResult(null, run.caseindex, null,"Transition ${problematicStep.fullLabel}", null, null, setOf(problematicStep.copy(labelParameters = HashMap(problematicStep.labelParameters), notes = HashSet(problematicStep.notes))))
-	}
+    override fun submitTest(run: TestCase): AnalysisResult {
+        val problematicStep = run.steps.sortedBy { it.stepNumber }.last()
+        return AnalysisResult(null, run.caseindex, null, "Transition ${problematicStep.fullLabel}", null, null, setOf(problematicStep.copy(labelParameters = HashMap(problematicStep.labelParameters), notes = HashSet(problematicStep.notes))))
+    }
 
 }
